requirements.txt
pid==3.0.3
psutil==5.7.0
requests>=2.20.0
python_logstash_async==1.6.4
confuse==1.1.0
pyzmq==19.0.1
webhook_listener>=0.0.3
